feat(badge): refactor hot badge item
